refactor: 의존성을 @ 기반 절대 경로로 표현하도록 변경, 폴더 구조 변경, parser / serializer 분리를 위해 converter을 다시 parser로 이름 수정
